update unit tests
