+++ src/AzureIoTHub.Portal.Tests.Unit/Client/Components/Devices/EditDeviceTests.cs
@@ -12,6 +12,7 @@ namespace AzureIoTHub.Portal.Tests.Unit.Client.Components.Devices
     using AzureIoTHub.Portal.Client.Enums;
     using AzureIoTHub.Portal.Client.Exceptions;
     using AzureIoTHub.Portal.Client.Models;
+    using AzureIoTHub.Portal.Client.Pages.Devices;
     using AzureIoTHub.Portal.Client.Services;
     using AzureIoTHub.Portal.Models.v10;
     using AzureIoTHub.Portal.Models.v10.LoRaWAN;
@@ -33,27 +34,28 @@ namespace AzureIoTHub.Portal.Tests.Unit.Client.Components.Devices
     public class EditDeviceTests : BlazorUnitTest
     {
         private Mock<IDialogService> mockDialogService;
+        //private Mock<ISnackbar> mockSnackbarService;
         private FakeNavigationManager mockNavigationManager;
         private Mock<IDeviceModelsClientService> mockDeviceModelsClientService;
         private Mock<ILoRaWanDeviceModelsClientService> mockLoRaWanDeviceModelsClientService;
         private Mock<IDeviceTagSettingsClientService> mockDeviceTagSettingsClientService;
         private Mock<IDeviceClientService> mockDeviceClientService;
         private Mock<ILoRaWanDeviceClientService> mockLoRaWanDeviceClientService;
-        private Mock<IDeviceLayoutService> mockDeviceLayoutService;
 
         public override void Setup()
         {
             base.Setup();
 
             this.mockDialogService = MockRepository.Create<IDialogService>();
+            //this.mockSnackbarService = MockRepository.Create<ISnackbar>();
             this.mockDeviceModelsClientService = MockRepository.Create<IDeviceModelsClientService>();
             this.mockLoRaWanDeviceModelsClientService = MockRepository.Create<ILoRaWanDeviceModelsClientService>();
             this.mockDeviceTagSettingsClientService = MockRepository.Create<IDeviceTagSettingsClientService>();
             this.mockDeviceClientService = MockRepository.Create<IDeviceClientService>();
             this.mockLoRaWanDeviceClientService = MockRepository.Create<ILoRaWanDeviceClientService>();
-            this.mockDeviceLayoutService = MockRepository.Create<IDeviceLayoutService>();
 
             _ = Services.AddSingleton(this.mockDialogService.Object);
+            //_ = Services.AddSingleton(this.mockSnackbarService.Object);
             _ = Services.AddSingleton(this.mockDeviceModelsClientService.Object);
             _ = Services.AddSingleton(this.mockLoRaWanDeviceModelsClientService.Object);
             _ = Services.AddSingleton(this.mockDeviceTagSettingsClientService.Object);
@@ -318,7 +320,7 @@ namespace AzureIoTHub.Portal.Tests.Unit.Client.Components.Devices
                 .Returns(Task.CompletedTask);
 
             var popoverProvider = RenderComponent<MudPopoverProvider>();
-            var cut = RenderComponent<EditDevice>();
+            var cut = RenderComponent<EditDevice>(parameters => parameters.Add(p => p.context, CreateEditMode.Create));
             var saveButton = cut.WaitForElement("#SaveButton");
 
             cut.WaitForElement($"#{nameof(DeviceDetails.DeviceName)}").Change(expectedDeviceDetails.DeviceName);
@@ -758,6 +760,559 @@ namespace AzureIoTHub.Portal.Tests.Unit.Client.Components.Devices
 
             // Assert
             cut.WaitForAssertion(() => cut.Find("div.mud-tooltip-root:nth-child(2) > div.mud-tab").TextContent.Should().Be("LoRaWAN"));
+        }
+
+        [Test]
+        public void ShouldLoadDeviceDetails()
+        {
+            // Arrange
+            var deviceId = Guid.NewGuid().ToString();
+            var modelId = Guid.NewGuid().ToString();
+
+            _ = this.mockDeviceClientService
+                .Setup(service => service.GetDevice(deviceId))
+                .ReturnsAsync(new DeviceDetails() { ModelId = modelId });
+
+            _ = this.mockDeviceClientService
+                .Setup(service => service.GetDeviceProperties(deviceId))
+                .ReturnsAsync(new List<DevicePropertyValue>());
+
+            _ = this.mockDeviceModelsClientService.Setup(service => service.GetDeviceModel(modelId))
+                .ReturnsAsync(new DeviceModelDto());
+
+            _ = this.mockDeviceTagSettingsClientService.Setup(service => service.GetDeviceTags())
+                .ReturnsAsync(new List<DeviceTagDto>());
+
+            // Act
+            var page = RenderComponent<DeviceDetailPage>(parameters => parameters.Add(p => p.DeviceID, deviceId));
+            var cut = RenderComponent<EditDevice>(parameters => parameters.Add(p => p.context, CreateEditMode.Edit).Add(p => p.DeviceID, deviceId));
+
+            // Assert
+            _ = page.WaitForElement("#returnButton");
+            cut.WaitForAssertion(() => MockRepository.VerifyAll());
+        }
+
+        [Test]
+        public void ShouldLoadLoRaDeviceDetails()
+        {
+
+            // Arrange
+            var deviceId = Guid.NewGuid().ToString();
+            var modelId = Guid.NewGuid().ToString();
+
+            _ = this.mockDeviceClientService
+                .Setup(service => service.GetDevice(deviceId))
+                .ReturnsAsync(new DeviceDetails());
+
+            _ = this.mockDeviceClientService
+                .Setup(service => service.GetDeviceProperties(deviceId))
+                .ReturnsAsync(new List<DevicePropertyValue>());
+
+            _ = this.mockDeviceModelsClientService.Setup(service => service.GetDeviceModel(null))
+                .ReturnsAsync(new DeviceModelDto());
+
+            _ = this.mockLoRaWanDeviceClientService
+                .Setup(service => service.GetDevice(deviceId))
+                .ReturnsAsync(new LoRaDeviceDetails() { ModelId = modelId });
+
+            _ = this.mockLoRaWanDeviceModelsClientService.Setup(service => service.GetDeviceModel(modelId))
+                .ReturnsAsync(new LoRaDeviceModelDto());
+
+            _ = this.mockLoRaWanDeviceModelsClientService.Setup(service => service.GetDeviceModelCommands(modelId))
+                .ReturnsAsync(new List<DeviceModelCommandDto>());
+
+            _ = this.mockDeviceTagSettingsClientService.Setup(service => service.GetDeviceTags())
+                .ReturnsAsync(new List<DeviceTagDto>());
+
+            // Act
+            var page = RenderComponent<DeviceDetailPage>(parameters => parameters.Add(p => p.DeviceID, deviceId));
+            var cut = RenderComponent<EditDevice>(parameters => parameters.Add(p => p.context, CreateEditMode.Edit).Add(p => p.DeviceID, deviceId).Add(p => p.IsLoRa, true));
+
+            // Assert
+            _ = page.WaitForElement("#returnButton");
+            cut.WaitForAssertion(() => MockRepository.VerifyAll());
+        }
+
+        [Test]
+        public void ReturnButtonMustNavigateToPreviousPage()
+        {
+
+            // Arrange
+            var deviceId = Guid.NewGuid().ToString();
+            var modelId = Guid.NewGuid().ToString();
+
+            _ = this.mockDeviceClientService
+                .Setup(service => service.GetDevice(deviceId))
+                .ReturnsAsync(new DeviceDetails() { ModelId = modelId });
+
+            _ = this.mockDeviceClientService
+                .Setup(service => service.GetDeviceProperties(deviceId))
+                .ReturnsAsync(new List<DevicePropertyValue>());
+
+            _ = this.mockDeviceModelsClientService.Setup(service => service.GetDeviceModel(modelId))
+                .ReturnsAsync(new DeviceModelDto());
+
+            _ = this.mockDeviceTagSettingsClientService.Setup(service => service.GetDeviceTags())
+                .ReturnsAsync(new List<DeviceTagDto>());
+
+
+            // Act
+            var page = RenderComponent<DeviceDetailPage>(parameters => parameters.Add(p => p.DeviceID, deviceId));
+            var cut = RenderComponent<EditDevice>(parameters => parameters.Add(p => p.context, CreateEditMode.Edit).Add(p => p.DeviceID, deviceId));
+            var returnButton = page.WaitForElement("#returnButton");
+
+            returnButton.Click();
+
+            // Assert
+            page.WaitForAssertion(() => this.mockNavigationManager.Uri.Should().EndWith("/devices"));
+            cut.WaitForAssertion(() => MockRepository.VerifyAll());
+        }
+
+        //[Test]
+        //public void ClickOnSaveShouldPutDeviceDetails()
+        //{
+        //    var mockDeviceModel = new DeviceModelDto
+        //    {
+        //        ModelId = Guid.NewGuid().ToString(),
+        //        Description = Guid.NewGuid().ToString(),
+        //        SupportLoRaFeatures = false,
+        //        Name = Guid.NewGuid().ToString()
+        //    };
+
+        //    var mockTag = new DeviceTagDto
+        //    {
+        //        Label = Guid.NewGuid().ToString(),
+        //        Name = Guid.NewGuid().ToString(),
+        //        Required = false,
+        //        Searchable = false
+        //    };
+
+        //    var mockDeviceDetails = new DeviceDetails
+        //    {
+        //        DeviceName = Guid.NewGuid().ToString(),
+        //        ModelId = mockDeviceModel.ModelId,
+        //        DeviceID = Guid.NewGuid().ToString(),
+        //        Tags = new Dictionary<string, string>()
+        //        {
+        //            {mockTag.Name,Guid.NewGuid().ToString()}
+        //        }
+        //    };
+
+        //    _ = this.mockDeviceClientService.Setup(service => service.UpdateDevice(It.Is<DeviceDetails>(details => mockDeviceDetails.DeviceID.Equals(details.DeviceID, StringComparison.Ordinal))))
+        //        .Returns(Task.CompletedTask);
+
+        //    _ = this.mockDeviceClientService
+        //        .Setup(service => service.GetDevice(mockDeviceDetails.DeviceID))
+        //        .ReturnsAsync(mockDeviceDetails);
+
+        //    _ = this.mockDeviceModelsClientService.Setup(service => service.GetDeviceModel(mockDeviceDetails.ModelId))
+        //        .ReturnsAsync(mockDeviceModel);
+
+        //    _ = this.mockDeviceTagSettingsClientService.Setup(service => service.GetDeviceTags())
+        //        .ReturnsAsync(new List<DeviceTagDto>
+        //        {
+        //            mockTag
+        //        });
+
+        //    _ = this.mockDeviceClientService
+        //        .Setup(service => service.GetDeviceProperties(mockDeviceDetails.DeviceID))
+        //        .ReturnsAsync(new List<DevicePropertyValue>());
+
+        //    _ = this.mockDeviceClientService
+        //        .Setup(service => service.SetDeviceProperties(mockDeviceDetails.DeviceID, It.IsAny<IList<DevicePropertyValue>>()))
+        //        .Returns(Task.CompletedTask);
+
+        //    _ = this.mockSnackbarService.Setup(c => c.Add(It.IsAny<string>(), Severity.Success, It.IsAny<Action<SnackbarOptions>>(), It.IsAny<string>())).Returns((Snackbar)null);
+
+        //    // Act
+        //    var cut = RenderComponent<EditDevice>(parameters => parameters.Add(p => p.context, CreateEditMode.Edit).Add(p => p.DeviceID, mockDeviceDetails.DeviceID));
+        //    cut.WaitForAssertion(() => cut.Find($"#{nameof(DeviceModelDto.Name)}").InnerHtml.Should().NotBeEmpty());
+
+        //    var saveButton = cut.WaitForElement("#saveButton");
+        //    saveButton.Click();
+
+        //    // Assert
+        //    cut.WaitForState(() => this.mockNavigationManager.Uri.EndsWith("devices", StringComparison.OrdinalIgnoreCase), 3.Seconds());
+        //    cut.WaitForAssertion(() => MockRepository.VerifyAll());
+        //}
+
+        [Test]
+        public void SaveShouldProcessProblemDetailsExceptionWhenIssueOccursOnUpdatingDevice()
+        {
+            var mockDeviceModel = new DeviceModelDto
+            {
+                ModelId = Guid.NewGuid().ToString(),
+                Description = Guid.NewGuid().ToString(),
+                SupportLoRaFeatures = false,
+                Name = Guid.NewGuid().ToString()
+            };
+
+            var mockTag = new DeviceTagDto
+            {
+                Label = Guid.NewGuid().ToString(),
+                Name = Guid.NewGuid().ToString(),
+                Required = false,
+                Searchable = false
+            };
+
+            var mockDeviceDetails = new DeviceDetails
+            {
+                DeviceName = Guid.NewGuid().ToString(),
+                ModelId = mockDeviceModel.ModelId,
+                DeviceID = Guid.NewGuid().ToString(),
+                Tags = new Dictionary<string, string>()
+                {
+                    {mockTag.Name,Guid.NewGuid().ToString()}
+                }
+            };
+
+            _ = this.mockDeviceClientService.Setup(service => service.UpdateDevice(It.Is<DeviceDetails>(details => mockDeviceDetails.DeviceID.Equals(details.DeviceID, StringComparison.Ordinal))))
+                .ThrowsAsync(new ProblemDetailsException(new ProblemDetailsWithExceptionDetails()));
+
+            _ = this.mockDeviceClientService
+                .Setup(service => service.GetDevice(mockDeviceDetails.DeviceID))
+                .ReturnsAsync(mockDeviceDetails);
+
+            _ = this.mockDeviceModelsClientService.Setup(service => service.GetDeviceModel(mockDeviceDetails.ModelId))
+                .ReturnsAsync(mockDeviceModel);
+
+            _ = this.mockDeviceTagSettingsClientService.Setup(service => service.GetDeviceTags())
+                .ReturnsAsync(new List<DeviceTagDto>
+                {
+                    mockTag
+                });
+
+            _ = this.mockDeviceClientService
+                .Setup(service => service.GetDeviceProperties(mockDeviceDetails.DeviceID))
+                .ReturnsAsync(new List<DevicePropertyValue>());
+
+            // Act
+            var cut = RenderComponent<EditDevice>(parameters => parameters.Add(p => p.context, CreateEditMode.Edit).Add(p => p.DeviceID, mockDeviceDetails.DeviceID));
+
+
+            var saveButton = cut.WaitForElement("#saveButton");
+            saveButton.Click();
+
+            // Assert
+            cut.WaitForAssertion(() => this.mockNavigationManager.Uri.Should().NotEndWith("devices"));
+            cut.WaitForAssertion(() => MockRepository.VerifyAll());
+        }
+
+        //[Test]
+        //public void ClickOnSaveShouldDisplaySnackbarIfValidationError()
+        //{
+        //    var mockDeviceModel = new DeviceModelDto
+        //    {
+        //        ModelId = Guid.NewGuid().ToString(),
+        //        Description = Guid.NewGuid().ToString(),
+        //        SupportLoRaFeatures = false,
+        //        Name = Guid.NewGuid().ToString()
+        //    };
+
+        //    var mockTag = new DeviceTagDto
+        //    {
+        //        Label = Guid.NewGuid().ToString(),
+        //        Name = Guid.NewGuid().ToString(),
+        //        Required = false,
+        //        Searchable = false
+        //    };
+
+        //    var mockDeviceDetails = new DeviceDetails
+        //    {
+        //        DeviceName = Guid.NewGuid().ToString(),
+        //        ModelId = mockDeviceModel.ModelId,
+        //        DeviceID = Guid.NewGuid().ToString(),
+        //        Tags = new Dictionary<string, string>()
+        //        {
+        //            {mockTag.Name,Guid.NewGuid().ToString()}
+        //        }
+        //    };
+
+        //    _ = this.mockDeviceClientService
+        //        .Setup(service => service.GetDevice(mockDeviceDetails.DeviceID))
+        //        .ReturnsAsync(mockDeviceDetails);
+
+        //    _ = this.mockDeviceModelsClientService.Setup(service => service.GetDeviceModel(mockDeviceDetails.ModelId))
+        //        .ReturnsAsync(mockDeviceModel);
+
+        //    _ = this.mockDeviceTagSettingsClientService.Setup(service => service.GetDeviceTags())
+        //        .ReturnsAsync(new List<DeviceTagDto>
+        //        {
+        //            mockTag
+        //        });
+
+        //    _ = this.mockDeviceClientService
+        //        .Setup(service => service.GetDeviceProperties(mockDeviceDetails.DeviceID))
+        //        .ReturnsAsync(new List<DevicePropertyValue>());
+
+        //    _ = this.mockSnackbarService.Setup(c => c.Add(It.IsAny<string>(), Severity.Error, It.IsAny<Action<SnackbarOptions>>(), It.IsAny<string>())).Returns((Snackbar)null);
+
+        //    // Act
+        //    var cut = RenderComponent<EditDevice>(parameters => parameters.Add(p => p.context, CreateEditMode.Edit).Add(p => p.DeviceID, mockDeviceDetails.DeviceID));
+
+        //    cut.WaitForElement($"#{nameof(DeviceDetails.DeviceName)}").Change("");
+        //    var saveButton = cut.WaitForElement("#saveButton");
+        //    saveButton.Click();
+
+        //    // Assert
+        //    cut.WaitForAssertion(() => MockRepository.VerifyAll());
+        //}
+
+        [Test]
+        public void ClickOnConnectShouldDisplayDeviceCredentials()
+        {
+            var mockDeviceModel = new DeviceModelDto
+            {
+                ModelId = Guid.NewGuid().ToString(),
+                Description = Guid.NewGuid().ToString(),
+                SupportLoRaFeatures = false,
+                Name = Guid.NewGuid().ToString()
+            };
+
+            var mockTag = new DeviceTagDto
+            {
+                Label = Guid.NewGuid().ToString(),
+                Name = Guid.NewGuid().ToString(),
+                Required = false,
+                Searchable = false
+            };
+
+            var mockDeviceDetails = new DeviceDetails
+            {
+                DeviceName = Guid.NewGuid().ToString(),
+                ModelId = mockDeviceModel.ModelId,
+                DeviceID = Guid.NewGuid().ToString(),
+                Tags = new Dictionary<string, string>()
+                {
+                    {mockTag.Name,Guid.NewGuid().ToString()}
+                }
+            };
+
+            _ = this.mockDeviceClientService
+                .Setup(service => service.GetDevice(mockDeviceDetails.DeviceID))
+                .ReturnsAsync(mockDeviceDetails);
+
+            _ = this.mockDeviceModelsClientService.Setup(service => service.GetDeviceModel(mockDeviceDetails.ModelId))
+                .ReturnsAsync(mockDeviceModel);
+
+            _ = this.mockDeviceTagSettingsClientService.Setup(service => service.GetDeviceTags())
+                .ReturnsAsync(new List<DeviceTagDto>
+                {
+                    mockTag
+                });
+
+            _ = this.mockDeviceClientService
+                .Setup(service => service.GetDeviceProperties(mockDeviceDetails.DeviceID))
+                .ReturnsAsync(new List<DevicePropertyValue>());
+
+            var mockDialogReference = new DialogReference(Guid.NewGuid(), this.mockDialogService.Object);
+            _ = this.mockDialogService.Setup(c => c.Show<ConnectionStringDialog>(It.IsAny<string>(), It.IsAny<DialogParameters>()))
+                .Returns(mockDialogReference);
+
+            // Act
+            var cut = RenderComponent<EditDevice>(parameters => parameters.Add(p => p.context, CreateEditMode.Edit).Add(p => p.DeviceID, mockDeviceDetails.DeviceID));
+
+            var connectButton = cut.WaitForElement("#connectButton");
+            connectButton.Click();
+
+            // Assert
+            cut.WaitForAssertion(() => MockRepository.VerifyAll());
+        }
+
+        [Test]
+        public void ClickOnDeleteShouldDisplayConfirmationDialogAndReturnIfAborted()
+        {
+            var mockDeviceModel = new DeviceModelDto
+            {
+                ModelId = Guid.NewGuid().ToString(),
+                Description = Guid.NewGuid().ToString(),
+                SupportLoRaFeatures = false,
+                Name = Guid.NewGuid().ToString()
+            };
+
+            var mockTag = new DeviceTagDto
+            {
+                Label = Guid.NewGuid().ToString(),
+                Name = Guid.NewGuid().ToString(),
+                Required = false,
+                Searchable = false
+            };
+
+            var mockDeviceDetails = new DeviceDetails
+            {
+                DeviceName = Guid.NewGuid().ToString(),
+                ModelId = mockDeviceModel.ModelId,
+                DeviceID = Guid.NewGuid().ToString(),
+                Tags = new Dictionary<string, string>()
+                {
+                    {mockTag.Name,Guid.NewGuid().ToString()}
+                }
+            };
+
+            _ = this.mockDeviceClientService
+                .Setup(service => service.GetDevice(mockDeviceDetails.DeviceID))
+                .ReturnsAsync(mockDeviceDetails);
+
+            _ = this.mockDeviceModelsClientService.Setup(service => service.GetDeviceModel(mockDeviceDetails.ModelId))
+                .ReturnsAsync(mockDeviceModel);
+
+            _ = this.mockDeviceTagSettingsClientService.Setup(service => service.GetDeviceTags())
+                .ReturnsAsync(new List<DeviceTagDto>
+                {
+                    mockTag
+                });
+
+            _ = this.mockDeviceClientService
+                .Setup(service => service.GetDeviceProperties(mockDeviceDetails.DeviceID))
+                .ReturnsAsync(new List<DevicePropertyValue>());
+
+            var mockDialogReference = MockRepository.Create<IDialogReference>();
+            _ = mockDialogReference.Setup(c => c.Result).ReturnsAsync(DialogResult.Cancel());
+            _ = this.mockDialogService.Setup(c => c.Show<DeleteDevicePage>(It.IsAny<string>(), It.IsAny<DialogParameters>()))
+                .Returns(mockDialogReference.Object);
+
+            // Act
+            var cut = RenderComponent<EditDevice>(parameters => parameters.Add(p => p.context, CreateEditMode.Edit).Add(p => p.DeviceID, mockDeviceDetails.DeviceID));
+
+            var deleteButton = cut.WaitForElement("#deleteButton");
+            deleteButton.Click();
+
+            // Assert
+            cut.WaitForAssertion(() => MockRepository.VerifyAll());
+        }
+
+        [Test]
+        public void ClickOnDeleteShouldDisplayConfirmationDialogAndRedirectIfConfirmed()
+        {
+            var mockDeviceModel = new DeviceModelDto
+            {
+                ModelId = Guid.NewGuid().ToString(),
+                Description = Guid.NewGuid().ToString(),
+                SupportLoRaFeatures = false,
+                Name = Guid.NewGuid().ToString()
+            };
+
+            var mockTag = new DeviceTagDto
+            {
+                Label = Guid.NewGuid().ToString(),
+                Name = Guid.NewGuid().ToString(),
+                Required = false,
+                Searchable = false
+            };
+
+            var mockDeviceDetails = new DeviceDetails
+            {
+                DeviceName = Guid.NewGuid().ToString(),
+                ModelId = mockDeviceModel.ModelId,
+                DeviceID = Guid.NewGuid().ToString(),
+                Tags = new Dictionary<string, string>()
+                {
+                    {mockTag.Name,Guid.NewGuid().ToString()}
+                }
+            };
+
+            _ = this.mockDeviceClientService
+                .Setup(service => service.GetDevice(mockDeviceDetails.DeviceID))
+                .ReturnsAsync(mockDeviceDetails);
+
+            _ = this.mockDeviceModelsClientService.Setup(service => service.GetDeviceModel(mockDeviceDetails.ModelId))
+                .ReturnsAsync(mockDeviceModel);
+
+            _ = this.mockDeviceTagSettingsClientService.Setup(service => service.GetDeviceTags())
+                .ReturnsAsync(new List<DeviceTagDto>
+                {
+                    mockTag
+                });
+
+            _ = this.mockDeviceClientService
+                .Setup(service => service.GetDeviceProperties(mockDeviceDetails.DeviceID))
+                .ReturnsAsync(new List<DevicePropertyValue>());
+
+            var mockDialogReference = MockRepository.Create<IDialogReference>();
+            _ = mockDialogReference.Setup(c => c.Result).ReturnsAsync(DialogResult.Ok("Ok"));
+            _ = this.mockDialogService.Setup(c => c.Show<DeleteDevicePage>(It.IsAny<string>(), It.IsAny<DialogParameters>()))
+                .Returns(mockDialogReference.Object);
+
+            // Act
+            var cut = RenderComponent<EditDevice>(parameters => parameters.Add(p => p.context, CreateEditMode.Edit).Add(p => p.DeviceID, mockDeviceDetails.DeviceID));
+
+            var deleteButton = cut.WaitForElement("#deleteButton");
+            deleteButton.Click();
+
+            // Assert            
+            cut.WaitForState(() => this.mockNavigationManager.Uri.EndsWith("/devices", StringComparison.OrdinalIgnoreCase));
+            cut.WaitForAssertion(() => MockRepository.VerifyAll());
+        }
+
+        [Test]
+        public void ClickOnDuplicateShouldDuplicateDeviceDetailAndRedirectToCreateDevicePage()
+        {
+            var mockDeviceModel = new DeviceModelDto
+            {
+                ModelId = Guid.NewGuid().ToString(),
+                Description = Guid.NewGuid().ToString(),
+                SupportLoRaFeatures = false,
+                Name = Guid.NewGuid().ToString()
+            };
+
+            var mockTag = new DeviceTagDto
+            {
+                Label = Guid.NewGuid().ToString(),
+                Name = Guid.NewGuid().ToString(),
+                Required = false,
+                Searchable = false
+            };
+
+            var mockDeviceDetails = new DeviceDetails
+            {
+                DeviceName = Guid.NewGuid().ToString(),
+                ModelId = mockDeviceModel.ModelId,
+                DeviceID = Guid.NewGuid().ToString(),
+                Tags = new Dictionary<string, string>()
+                {
+                    {mockTag.Name,Guid.NewGuid().ToString()}
+                }
+            };
+
+            _ = this.mockDeviceClientService
+                .Setup(service => service.GetDevice(mockDeviceDetails.DeviceID))
+                .ReturnsAsync(mockDeviceDetails);
+
+            _ = this.mockDeviceModelsClientService.Setup(service => service.GetDeviceModel(mockDeviceDetails.ModelId))
+                .ReturnsAsync(mockDeviceModel);
+
+            _ = this.mockDeviceTagSettingsClientService.Setup(service => service.GetDeviceTags())
+                .ReturnsAsync(new List<DeviceTagDto>
+                {
+                    mockTag
+                });
+
+            _ = this.mockDeviceClientService
+                .Setup(service => service.GetDeviceProperties(mockDeviceDetails.DeviceID))
+                .ReturnsAsync(new List<DevicePropertyValue>());
+
+            var popoverProvider = RenderComponent<MudPopoverProvider>();
+            var cut = RenderComponent<EditDevice>(parameters => parameters.Add(p => p.context, CreateEditMode.Edit).Add(p => p.DeviceID, mockDeviceDetails.DeviceID));
+            cut.WaitForAssertion(() => cut.Find($"#{nameof(DeviceModelDto.Name)}").InnerHtml.Should().NotBeEmpty());
+
+            var saveButton = cut.WaitForElement("#saveButton");
+
+            var mudButtonGroup = cut.FindComponent<MudButtonGroup>();
+
+            mudButtonGroup.Find(".mud-menu button").Click();
+            popoverProvider.WaitForAssertion(() => popoverProvider.FindAll("div.mud-list-item").Count.Should().Be(2));
+
+            var items = popoverProvider.FindAll("div.mud-list-item");
+
+            // Click on Duplicate
+            items[1].Click();
+
+            // Act
+            saveButton.Click();
+
+            // Assert
+            cut.WaitForAssertion(() => MockRepository.VerifyAll());
+            cut.WaitForAssertion(() => this.mockNavigationManager.Uri.Should().EndWith("/devices/new"));
         }
     }
 }
