+++ src/AzureIoTHub.Portal.Tests.Unit/Client/Pages/Devices/DeviceDetailPageTests.cs
@@ -5,9 +5,6 @@ namespace AzureIoTHub.Portal.Tests.Unit.Client.Pages.Devices
 {
     using System;
     using System.Collections.Generic;
-    using System.Threading.Tasks;
-    using AzureIoTHub.Portal.Client.Exceptions;
-    using AzureIoTHub.Portal.Client.Models;
     using AzureIoTHub.Portal.Client.Pages.Devices;
     using AzureIoTHub.Portal.Client.Services;
     using Models.v10;
@@ -15,15 +12,12 @@ namespace AzureIoTHub.Portal.Tests.Unit.Client.Pages.Devices
     using Bunit;
     using Bunit.TestDoubles;
     using FluentAssertions;
-    using FluentAssertions.Extensions;
     using Microsoft.Extensions.DependencyInjection;
     using Moq;
     using MudBlazor;
     using MudBlazor.Services;
     using NUnit.Framework;
     using UnitTests.Mocks;
-    using Models.v10.LoRaWAN;
-    using AzureIoTHub.Portal.Client.Pages.DeviceModels;
 
     [TestFixture]
     public class DeviceDetailPageTests : BlazorUnitTest
@@ -68,541 +62,39 @@ namespace AzureIoTHub.Portal.Tests.Unit.Client.Pages.Devices
         }
 
         [Test]
-        public void ShouldLoadDeviceDetails()
+        public void DeviceDetailPageShouldRenderCorrectly()
         {
             // Arrange
             var deviceId = Guid.NewGuid().ToString();
-            var modelId = Guid.NewGuid().ToString();
+
+            _ = this.mockDeviceTagSettingsClientService.Setup(service => service.GetDeviceTags())
+                .ReturnsAsync(new List<DeviceTagDto>
+                {
+                    new()
+                    {
+                        Label = Guid.NewGuid().ToString(),
+                        Name = Guid.NewGuid().ToString(),
+                        Required = false,
+                        Searchable = false
+                    }
+                });
 
             _ = this.mockDeviceClientService
                 .Setup(service => service.GetDevice(deviceId))
-                .ReturnsAsync(new DeviceDetails() { ModelId = modelId });
+                .ReturnsAsync(new DeviceDetails());
 
             _ = this.mockDeviceClientService
                 .Setup(service => service.GetDeviceProperties(deviceId))
                 .ReturnsAsync(new List<DevicePropertyValue>());
 
-            _ = this.mockDeviceModelsClientService.Setup(service => service.GetDeviceModel(modelId))
+            _ = this.mockDeviceModelsClientService.Setup(service => service.GetDeviceModel(null))
                 .ReturnsAsync(new DeviceModelDto());
-
-            _ = this.mockDeviceTagSettingsClientService.Setup(service => service.GetDeviceTags())
-                .ReturnsAsync(new List<DeviceTagDto>());
 
             // Act
             var cut = RenderComponent<DeviceDetailPage>(ComponentParameter.CreateParameter("DeviceID", deviceId));
 
             // Assert
-            _ = cut.WaitForElement("#returnButton");
-        }
-
-        [Test]
-        public void ShouldLoadLoRaDeviceDetails()
-        {
-
-            // Arrange
-            var deviceId = Guid.NewGuid().ToString();
-            var modelId = Guid.NewGuid().ToString();
-
-            _ = this.mockLoRaWanDeviceClientService
-                .Setup(service => service.GetDevice(deviceId))
-                .ReturnsAsync(new LoRaDeviceDetails() { ModelId = modelId });
-
-            _ = this.mockLoRaWanDeviceModelsClientService.Setup(service => service.GetDeviceModel(modelId))
-                .ReturnsAsync(new LoRaDeviceModelDto());
-
-            _ = this.mockLoRaWanDeviceModelsClientService.Setup(service => service.GetDeviceModelCommands(modelId))
-                .ReturnsAsync(new List<DeviceModelCommandDto>());
-
-            _ = this.mockDeviceTagSettingsClientService.Setup(service => service.GetDeviceTags())
-                .ReturnsAsync(new List<DeviceTagDto>());
-
-            // Act
-            var cut = RenderComponent<DeviceDetailPage>(
-                    ComponentParameter.CreateParameter("DeviceID", deviceId),
-                    ComponentParameter.CreateParameter(nameof(DeviceModelDetailPage.IsLoRa), true));
-
-            // Assert
-            _ = cut.WaitForElement("#returnButton");
-        }
-
-        [Test]
-        public void ReturnButtonMustNavigateToPreviousPage()
-        {
-
-            // Arrange
-            var deviceId = Guid.NewGuid().ToString();
-            var modelId = Guid.NewGuid().ToString();
-
-            _ = this.mockDeviceClientService
-                .Setup(service => service.GetDevice(deviceId))
-                .ReturnsAsync(new DeviceDetails() { ModelId = modelId });
-
-            _ = this.mockDeviceClientService
-                .Setup(service => service.GetDeviceProperties(deviceId))
-                .ReturnsAsync(new List<DevicePropertyValue>());
-
-            _ = this.mockDeviceModelsClientService.Setup(service => service.GetDeviceModel(modelId))
-                .ReturnsAsync(new DeviceModelDto());
-
-            _ = this.mockDeviceTagSettingsClientService.Setup(service => service.GetDeviceTags())
-                .ReturnsAsync(new List<DeviceTagDto>());
-
-
-            // Act
-            var cut = RenderComponent<DeviceDetailPage>(ComponentParameter.CreateParameter("DeviceID", deviceId));
-            var returnButton = cut.WaitForElement("#returnButton");
-
-            returnButton.Click();
-
-            // Assert
-            cut.WaitForAssertion(() => this.mockNavigationManager.Uri.Should().EndWith("/devices"));
-            cut.WaitForAssertion(() => MockRepository.VerifyAll());
-        }
-
-        [Test]
-        public void ClickOnSaveShouldPutDeviceDetails()
-        {
-            var mockDeviceModel = new DeviceModelDto
-            {
-                ModelId = Guid.NewGuid().ToString(),
-                Description = Guid.NewGuid().ToString(),
-                SupportLoRaFeatures = false,
-                Name = Guid.NewGuid().ToString()
-            };
-
-            var mockTag = new DeviceTagDto
-            {
-                Label = Guid.NewGuid().ToString(),
-                Name = Guid.NewGuid().ToString(),
-                Required = false,
-                Searchable = false
-            };
-
-            var mockDeviceDetails = new DeviceDetails
-            {
-                DeviceName = Guid.NewGuid().ToString(),
-                ModelId = mockDeviceModel.ModelId,
-                DeviceID = Guid.NewGuid().ToString(),
-                Tags = new Dictionary<string, string>()
-                {
-                    {mockTag.Name,Guid.NewGuid().ToString()}
-                }
-            };
-
-            _ = this.mockDeviceClientService.Setup(service => service.UpdateDevice(It.Is<DeviceDetails>(details => mockDeviceDetails.DeviceID.Equals(details.DeviceID, StringComparison.Ordinal))))
-                .Returns(Task.CompletedTask);
-
-            _ = this.mockDeviceClientService
-                .Setup(service => service.GetDevice(mockDeviceDetails.DeviceID))
-                .ReturnsAsync(mockDeviceDetails);
-
-            _ = this.mockDeviceModelsClientService.Setup(service => service.GetDeviceModel(mockDeviceDetails.ModelId))
-                .ReturnsAsync(mockDeviceModel);
-
-            _ = this.mockDeviceTagSettingsClientService.Setup(service => service.GetDeviceTags())
-                .ReturnsAsync(new List<DeviceTagDto>
-                {
-                    mockTag
-                });
-
-            _ = this.mockDeviceClientService
-                .Setup(service => service.GetDeviceProperties(mockDeviceDetails.DeviceID))
-                .ReturnsAsync(new List<DevicePropertyValue>());
-
-            _ = this.mockDeviceClientService
-                .Setup(service => service.SetDeviceProperties(mockDeviceDetails.DeviceID, It.IsAny<IList<DevicePropertyValue>>()))
-                .Returns(Task.CompletedTask);
-
-            _ = this.mockSnackbarService.Setup(c => c.Add(It.IsAny<string>(), Severity.Success, It.IsAny<Action<SnackbarOptions>>(), It.IsAny<string>())).Returns((Snackbar)null);
-
-            // Act
-            var cut = RenderComponent<DeviceDetailPage>(ComponentParameter.CreateParameter("DeviceID", mockDeviceDetails.DeviceID));
-            cut.WaitForAssertion(() => cut.Find($"#{nameof(DeviceModelDto.Name)}").InnerHtml.Should().NotBeEmpty());
-
-            var saveButton = cut.WaitForElement("#saveButton");
-            saveButton.Click();
-
-            // Assert
-            cut.WaitForState(() => this.mockNavigationManager.Uri.EndsWith("devices", StringComparison.OrdinalIgnoreCase), 3.Seconds());
-            cut.WaitForAssertion(() => MockRepository.VerifyAll());
-        }
-
-        [Test]
-        public void SaveShouldProcessProblemDetailsExceptionWhenIssueOccursOnUpdatingDevice()
-        {
-            var mockDeviceModel = new DeviceModelDto
-            {
-                ModelId = Guid.NewGuid().ToString(),
-                Description = Guid.NewGuid().ToString(),
-                SupportLoRaFeatures = false,
-                Name = Guid.NewGuid().ToString()
-            };
-
-            var mockTag = new DeviceTagDto
-            {
-                Label = Guid.NewGuid().ToString(),
-                Name = Guid.NewGuid().ToString(),
-                Required = false,
-                Searchable = false
-            };
-
-            var mockDeviceDetails = new DeviceDetails
-            {
-                DeviceName = Guid.NewGuid().ToString(),
-                ModelId = mockDeviceModel.ModelId,
-                DeviceID = Guid.NewGuid().ToString(),
-                Tags = new Dictionary<string, string>()
-                {
-                    {mockTag.Name,Guid.NewGuid().ToString()}
-                }
-            };
-
-            _ = this.mockDeviceClientService.Setup(service => service.UpdateDevice(It.Is<DeviceDetails>(details => mockDeviceDetails.DeviceID.Equals(details.DeviceID, StringComparison.Ordinal))))
-                .ThrowsAsync(new ProblemDetailsException(new ProblemDetailsWithExceptionDetails()));
-
-            _ = this.mockDeviceClientService
-                .Setup(service => service.GetDevice(mockDeviceDetails.DeviceID))
-                .ReturnsAsync(mockDeviceDetails);
-
-            _ = this.mockDeviceModelsClientService.Setup(service => service.GetDeviceModel(mockDeviceDetails.ModelId))
-                .ReturnsAsync(mockDeviceModel);
-
-            _ = this.mockDeviceTagSettingsClientService.Setup(service => service.GetDeviceTags())
-                .ReturnsAsync(new List<DeviceTagDto>
-                {
-                    mockTag
-                });
-
-            _ = this.mockDeviceClientService
-                .Setup(service => service.GetDeviceProperties(mockDeviceDetails.DeviceID))
-                .ReturnsAsync(new List<DevicePropertyValue>());
-
-            // Act
-            var cut = RenderComponent<DeviceDetailPage>(ComponentParameter.CreateParameter("DeviceID", mockDeviceDetails.DeviceID));
-
-            var saveButton = cut.WaitForElement("#saveButton");
-            saveButton.Click();
-
-            // Assert
-            cut.WaitForAssertion(() => this.mockNavigationManager.Uri.Should().NotEndWith("devices"));
-            cut.WaitForAssertion(() => MockRepository.VerifyAll());
-        }
-
-        [Test]
-        public void ClickOnSaveShouldDisplaySnackbarIfValidationError()
-        {
-            var mockDeviceModel = new DeviceModelDto
-            {
-                ModelId = Guid.NewGuid().ToString(),
-                Description = Guid.NewGuid().ToString(),
-                SupportLoRaFeatures = false,
-                Name = Guid.NewGuid().ToString()
-            };
-
-            var mockTag = new DeviceTagDto
-            {
-                Label = Guid.NewGuid().ToString(),
-                Name = Guid.NewGuid().ToString(),
-                Required = false,
-                Searchable = false
-            };
-
-            var mockDeviceDetails = new DeviceDetails
-            {
-                DeviceName = Guid.NewGuid().ToString(),
-                ModelId = mockDeviceModel.ModelId,
-                DeviceID = Guid.NewGuid().ToString(),
-                Tags = new Dictionary<string, string>()
-                {
-                    {mockTag.Name,Guid.NewGuid().ToString()}
-                }
-            };
-
-            _ = this.mockDeviceClientService
-                .Setup(service => service.GetDevice(mockDeviceDetails.DeviceID))
-                .ReturnsAsync(mockDeviceDetails);
-
-            _ = this.mockDeviceModelsClientService.Setup(service => service.GetDeviceModel(mockDeviceDetails.ModelId))
-                .ReturnsAsync(mockDeviceModel);
-
-            _ = this.mockDeviceTagSettingsClientService.Setup(service => service.GetDeviceTags())
-                .ReturnsAsync(new List<DeviceTagDto>
-                {
-                    mockTag
-                });
-
-            _ = this.mockDeviceClientService
-                .Setup(service => service.GetDeviceProperties(mockDeviceDetails.DeviceID))
-                .ReturnsAsync(new List<DevicePropertyValue>());
-
-            _ = this.mockSnackbarService.Setup(c => c.Add(It.IsAny<string>(), Severity.Error, It.IsAny<Action<SnackbarOptions>>(), It.IsAny<string>())).Returns((Snackbar)null);
-
-            // Act
-            var cut = RenderComponent<DeviceDetailPage>(ComponentParameter.CreateParameter("DeviceID", mockDeviceDetails.DeviceID));
-
-            cut.WaitForElement($"#{nameof(DeviceDetails.DeviceName)}").Change("");
-            var saveButton = cut.WaitForElement("#saveButton");
-            saveButton.Click();
-
-            // Assert
-            cut.WaitForAssertion(() => MockRepository.VerifyAll());
-        }
-
-        [Test]
-        public void ClickOnConnectShouldDisplayDeviceCredentials()
-        {
-            var mockDeviceModel = new DeviceModelDto
-            {
-                ModelId = Guid.NewGuid().ToString(),
-                Description = Guid.NewGuid().ToString(),
-                SupportLoRaFeatures = false,
-                Name = Guid.NewGuid().ToString()
-            };
-
-            var mockTag = new DeviceTagDto
-            {
-                Label = Guid.NewGuid().ToString(),
-                Name = Guid.NewGuid().ToString(),
-                Required = false,
-                Searchable = false
-            };
-
-            var mockDeviceDetails = new DeviceDetails
-            {
-                DeviceName = Guid.NewGuid().ToString(),
-                ModelId = mockDeviceModel.ModelId,
-                DeviceID = Guid.NewGuid().ToString(),
-                Tags = new Dictionary<string, string>()
-                {
-                    {mockTag.Name,Guid.NewGuid().ToString()}
-                }
-            };
-
-            _ = this.mockDeviceClientService
-                .Setup(service => service.GetDevice(mockDeviceDetails.DeviceID))
-                .ReturnsAsync(mockDeviceDetails);
-
-            _ = this.mockDeviceModelsClientService.Setup(service => service.GetDeviceModel(mockDeviceDetails.ModelId))
-                .ReturnsAsync(mockDeviceModel);
-
-            _ = this.mockDeviceTagSettingsClientService.Setup(service => service.GetDeviceTags())
-                .ReturnsAsync(new List<DeviceTagDto>
-                {
-                    mockTag
-                });
-
-            _ = this.mockDeviceClientService
-                .Setup(service => service.GetDeviceProperties(mockDeviceDetails.DeviceID))
-                .ReturnsAsync(new List<DevicePropertyValue>());
-
-            var mockDialogReference = new DialogReference(Guid.NewGuid(), this.mockDialogService.Object);
-            _ = this.mockDialogService.Setup(c => c.Show<ConnectionStringDialog>(It.IsAny<string>(), It.IsAny<DialogParameters>()))
-                .Returns(mockDialogReference);
-
-            // Act
-            var cut = RenderComponent<DeviceDetailPage>(ComponentParameter.CreateParameter("DeviceID", mockDeviceDetails.DeviceID));
-
-            var connectButton = cut.WaitForElement("#connectButton");
-            connectButton.Click();
-
-            // Assert
-            cut.WaitForAssertion(() => MockRepository.VerifyAll());
-        }
-
-        [Test]
-        public void ClickOnDeleteShouldDisplayConfirmationDialogAndReturnIfAborted()
-        {
-            var mockDeviceModel = new DeviceModelDto
-            {
-                ModelId = Guid.NewGuid().ToString(),
-                Description = Guid.NewGuid().ToString(),
-                SupportLoRaFeatures = false,
-                Name = Guid.NewGuid().ToString()
-            };
-
-            var mockTag = new DeviceTagDto
-            {
-                Label = Guid.NewGuid().ToString(),
-                Name = Guid.NewGuid().ToString(),
-                Required = false,
-                Searchable = false
-            };
-
-            var mockDeviceDetails = new DeviceDetails
-            {
-                DeviceName = Guid.NewGuid().ToString(),
-                ModelId = mockDeviceModel.ModelId,
-                DeviceID = Guid.NewGuid().ToString(),
-                Tags = new Dictionary<string, string>()
-                {
-                    {mockTag.Name,Guid.NewGuid().ToString()}
-                }
-            };
-
-            _ = this.mockDeviceClientService
-                .Setup(service => service.GetDevice(mockDeviceDetails.DeviceID))
-                .ReturnsAsync(mockDeviceDetails);
-
-            _ = this.mockDeviceModelsClientService.Setup(service => service.GetDeviceModel(mockDeviceDetails.ModelId))
-                .ReturnsAsync(mockDeviceModel);
-
-            _ = this.mockDeviceTagSettingsClientService.Setup(service => service.GetDeviceTags())
-                .ReturnsAsync(new List<DeviceTagDto>
-                {
-                    mockTag
-                });
-
-            _ = this.mockDeviceClientService
-                .Setup(service => service.GetDeviceProperties(mockDeviceDetails.DeviceID))
-                .ReturnsAsync(new List<DevicePropertyValue>());
-
-            var mockDialogReference = MockRepository.Create<IDialogReference>();
-            _ = mockDialogReference.Setup(c => c.Result).ReturnsAsync(DialogResult.Cancel());
-            _ = this.mockDialogService.Setup(c => c.Show<DeleteDevicePage>(It.IsAny<string>(), It.IsAny<DialogParameters>()))
-                .Returns(mockDialogReference.Object);
-
-            // Act
-            var cut = RenderComponent<DeviceDetailPage>(ComponentParameter.CreateParameter("DeviceID", mockDeviceDetails.DeviceID));
-
-            var deleteButton = cut.WaitForElement("#deleteButton");
-            deleteButton.Click();
-
-            // Assert
-            cut.WaitForAssertion(() => MockRepository.VerifyAll());
-        }
-
-        [Test]
-        public void ClickOnDeleteShouldDisplayConfirmationDialogAndRedirectIfConfirmed()
-        {
-            var mockDeviceModel = new DeviceModelDto
-            {
-                ModelId = Guid.NewGuid().ToString(),
-                Description = Guid.NewGuid().ToString(),
-                SupportLoRaFeatures = false,
-                Name = Guid.NewGuid().ToString()
-            };
-
-            var mockTag = new DeviceTagDto
-            {
-                Label = Guid.NewGuid().ToString(),
-                Name = Guid.NewGuid().ToString(),
-                Required = false,
-                Searchable = false
-            };
-
-            var mockDeviceDetails = new DeviceDetails
-            {
-                DeviceName = Guid.NewGuid().ToString(),
-                ModelId = mockDeviceModel.ModelId,
-                DeviceID = Guid.NewGuid().ToString(),
-                Tags = new Dictionary<string, string>()
-                {
-                    {mockTag.Name,Guid.NewGuid().ToString()}
-                }
-            };
-
-            _ = this.mockDeviceClientService
-                .Setup(service => service.GetDevice(mockDeviceDetails.DeviceID))
-                .ReturnsAsync(mockDeviceDetails);
-
-            _ = this.mockDeviceModelsClientService.Setup(service => service.GetDeviceModel(mockDeviceDetails.ModelId))
-                .ReturnsAsync(mockDeviceModel);
-
-            _ = this.mockDeviceTagSettingsClientService.Setup(service => service.GetDeviceTags())
-                .ReturnsAsync(new List<DeviceTagDto>
-                {
-                    mockTag
-                });
-
-            _ = this.mockDeviceClientService
-                .Setup(service => service.GetDeviceProperties(mockDeviceDetails.DeviceID))
-                .ReturnsAsync(new List<DevicePropertyValue>());
-
-            var mockDialogReference = MockRepository.Create<IDialogReference>();
-            _ = mockDialogReference.Setup(c => c.Result).ReturnsAsync(DialogResult.Ok("Ok"));
-            _ = this.mockDialogService.Setup(c => c.Show<DeleteDevicePage>(It.IsAny<string>(), It.IsAny<DialogParameters>()))
-                .Returns(mockDialogReference.Object);
-
-            // Act
-            var cut = RenderComponent<DeviceDetailPage>(ComponentParameter.CreateParameter("DeviceID", mockDeviceDetails.DeviceID));
-
-            var deleteButton = cut.WaitForElement("#deleteButton");
-            deleteButton.Click();
-
-            // Assert            
-            cut.WaitForState(() => this.mockNavigationManager.Uri.EndsWith("/devices", StringComparison.OrdinalIgnoreCase));
-            cut.WaitForAssertion(() => MockRepository.VerifyAll());
-        }
-
-        [Test]
-        public void ClickOnDuplicateShouldDuplicateDeviceDetailAndRedirectToCreateDevicePage()
-        {
-            var mockDeviceModel = new DeviceModelDto
-            {
-                ModelId = Guid.NewGuid().ToString(),
-                Description = Guid.NewGuid().ToString(),
-                SupportLoRaFeatures = false,
-                Name = Guid.NewGuid().ToString()
-            };
-
-            var mockTag = new DeviceTagDto
-            {
-                Label = Guid.NewGuid().ToString(),
-                Name = Guid.NewGuid().ToString(),
-                Required = false,
-                Searchable = false
-            };
-
-            var mockDeviceDetails = new DeviceDetails
-            {
-                DeviceName = Guid.NewGuid().ToString(),
-                ModelId = mockDeviceModel.ModelId,
-                DeviceID = Guid.NewGuid().ToString(),
-                Tags = new Dictionary<string, string>()
-                {
-                    {mockTag.Name,Guid.NewGuid().ToString()}
-                }
-            };
-
-            _ = this.mockDeviceClientService
-                .Setup(service => service.GetDevice(mockDeviceDetails.DeviceID))
-                .ReturnsAsync(mockDeviceDetails);
-
-            _ = this.mockDeviceModelsClientService.Setup(service => service.GetDeviceModel(mockDeviceDetails.ModelId))
-                .ReturnsAsync(mockDeviceModel);
-
-            _ = this.mockDeviceTagSettingsClientService.Setup(service => service.GetDeviceTags())
-                .ReturnsAsync(new List<DeviceTagDto>
-                {
-                    mockTag
-                });
-
-            _ = this.mockDeviceClientService
-                .Setup(service => service.GetDeviceProperties(mockDeviceDetails.DeviceID))
-                .ReturnsAsync(new List<DevicePropertyValue>());
-
-            var popoverProvider = RenderComponent<MudPopoverProvider>();
-            var cut = RenderComponent<DeviceDetailPage>(ComponentParameter.CreateParameter("DeviceID", mockDeviceDetails.DeviceID));
-            cut.WaitForAssertion(() => cut.Find($"#{nameof(DeviceModelDto.Name)}").InnerHtml.Should().NotBeEmpty());
-
-            var saveButton = cut.WaitForElement("#saveButton");
-
-            var mudButtonGroup = cut.FindComponent<MudButtonGroup>();
-
-            mudButtonGroup.Find(".mud-menu button").Click();
-            popoverProvider.WaitForAssertion(() => popoverProvider.FindAll("div.mud-list-item").Count.Should().Be(2));
-
-            var items = popoverProvider.FindAll("div.mud-list-item");
-
-            // Click on Duplicate
-            items[1].Click();
-
-            // Act
-            saveButton.Click();
-
-            // Assert
-            cut.WaitForAssertion(() => MockRepository.VerifyAll());
-            cut.WaitForAssertion(() => this.mockNavigationManager.Uri.Should().EndWith("/devices/new"));
+            cut.WaitForAssertion(() => cut.Find("#saveButton").TextContent.Should().Be("Save"));
         }
     }
 }
